\name{frame-package}
\alias{frame-package}
\docType{package}
\title{
The frame Package
}
\description{
Variables with Context
}
\details{
This package provides an extension to the `data.frame` type that makes it easy
to affix sampling frames to variables. It also provides a nice print function
than the `data.frame` default.

For a complete list of functions, use \code{library(help = "frame")}.
}
\author{
Patrick O. Perry
}
\keyword{ package }
