\name{keyrows}
\alias{keyrows}
\alias{keyrows.default}
\alias{keyrows.keyset}
\alias{keyrows.NULL}
\title{Data Key Rows}
\description{
Find the set of rows with specified key component values.
}
\usage{
keyrows(x, slice, ...)
}
\arguments{
\item{x}{an object with a key for each row.}

\item{slice}{the slice description, giving values for each key component.}

\item{...}{arguments specifying the slice.}
}
\details{
Extract a subset of \code{x} by getting the slice of rows with key
components belonging to the specified sets.

(fill in)
}
%\value{
%}
\seealso{
\code{\link{keys}}, \code{\link{[.dataset}}, \code{\link{rowid}}.
}
\examples{
# fill in
}
