\name{keyslice}
\alias{keyslice}
\alias{keyslice.default}
\alias{keyslice.keyset}
\title{Data Key Slicing}
\description{
Find sets of rows with specified key components.
}
\usage{
keyslice(x, ...)
}
\arguments{
\item{x}{an object with a key for each row.}

\item{...}{arguments specifying the slice.}
}
\details{
Extract a subset of \code{x} by getting the slice of rows with key
components belonging to the specified sets.

(fill in)
}
%\value{
%}
\seealso{
\code{\link{keys}}, \code{\link{[.dataset}}, \code{\link{rowid}}.
}
\examples{
# fill in
}
